Allow Concord environment creation to be overridden so that values such as the injector can be captured.
